Aztec3 Specific Work in Barretenberg (https://github.com/AztecProtocol/barretenberg/pull/142)
* Split Pedersen Hash & Commitment Gadgets (https://github.com/AztecProtocol/barretenberg/pull/95)

* [SQUASHED] Pedersen refactor into hash and commitment.

Use lookup pedersen for merkle tree, fixed-base pedersen for commitments.
---------

Port `copy_as_new_witness`.

Port `must_imply`.

`operator++`.

Port changes from `common`.

Port `ecc/groups`.

* [CPM] add missing dependencies to libbarretenberg.a (https://github.com/AztecProtocol/barretenberg/pull/154)
---------

* Increase Pedersen Generator indices and subindices. (https://github.com/AztecProtocol/barretenberg/pull/169)

* Remove a3 specific types. (https://github.com/AztecProtocol/barretenberg/pull/252)

* Address Luke's Comments on `aztec3 -> master` (https://github.com/AztecProtocol/barretenberg/pull/263)

* Add must_imply tests.

* Added a test for `field_t::copy_as_new_witness`

* add test for `conditional_assign`

* Added `infinity` test.

* Add `add_affine_test`.

* Tests for Array Object in `stdlib` (https://github.com/AztecProtocol/barretenberg/pull/262)

* basic array tests.

* Add `composer_type` while hashing/compressing a vkey.

* Add `contains_recursive_proof` to Recursive VK (https://github.com/AztecProtocol/barretenberg/pull/268)

* feat: debug utility for serialization (https://github.com/AztecProtocol/barretenberg/pull/290)

* feat: enable asan config

* `array_push` for Generic Type (https://github.com/AztecProtocol/barretenberg/pull/291)

* Add Indexed Merkle Tree  (https://github.com/AztecProtocol/barretenberg/pull/281)

* remove ts (consulted with Adam and we're good to go). (https://github.com/AztecProtocol/barretenberg/pull/292)

* Add cout for verification_key struct (https://github.com/AztecProtocol/barretenberg/pull/295)

* compute tree (https://github.com/AztecProtocol/barretenberg/pull/298)

* [SQUASHED] fixing `push_array_to_array` method. (https://github.com/AztecProtocol/barretenberg/pull/304)

* feat(memory_tree|a3): add sibling path calculations (https://github.com/AztecProtocol/barretenberg/pull/301)

* feat(memory_tree): frontier paths

* fix array and resolve merge conflicts (https://github.com/AztecProtocol/barretenberg/pull/305)

* Mc/hash vk (https://github.com/AztecProtocol/barretenberg/pull/306)

* Increase number of sub-generators to 128.

* Build a3crypto.wasm (https://github.com/AztecProtocol/barretenberg/pull/311)

* More Tests on A3 `stdlib` methods (https://github.com/AztecProtocol/barretenberg/pull/316)

* test: more vk tests to compare circuit/native/vk_data (https://github.com/AztecProtocol/barretenberg/pull/310)

* Mc/hash vk (https://github.com/AztecProtocol/barretenberg/pull/306)

* inc num_generators_per_hash_index to 128. (https://github.com/AztecProtocol/barretenberg/pull/309)

* fix. (https://github.com/AztecProtocol/barretenberg/pull/318)

* Added test for `compute_tree_native`. (https://github.com/AztecProtocol/barretenberg/pull/319)

* Install instructions for apt on ubuntu (https://github.com/AztecProtocol/barretenberg/pull/312)

* Fix address compilation. (https://github.com/AztecProtocol/barretenberg/pull/329)

---------

+++ circuits/cpp/barretenberg/cpp/src/barretenberg/stdlib/primitives/safe_uint/safe_uint.test.cpp
@@ -472,7 +472,8 @@ TEST(stdlib_safeuint, test_slice_random)
 
     const uint256_t expected0 = uint256_t(a_) & ((uint256_t(1) << uint64_t(lsb)) - 1);
     const uint256_t expected1 = (uint256_t(a_) >> lsb) & ((uint256_t(1) << (uint64_t(msb - lsb) + 1)) - 1);
-    const uint256_t expected2 = (uint256_t(a_) >> (msb + 1)) & ((uint256_t(1) << (uint64_t(252 - msb) - 1)) - 1);
+    const uint256_t expected2 =
+        (uint256_t(a_) >> uint64_t(msb + 1)) & ((uint256_t(1) << (uint64_t(252 - msb) - 1)) - 1);
 
     EXPECT_EQ(slice[0].get_value(), fr(expected0));
     EXPECT_EQ(slice[1].get_value(), fr(expected1));
